+++ packages/kit/test/apps/basics/src/routes/load/server-data-reuse/with-changing-parent/no-load/+page.svelte
@@ -1,0 +1,7 @@
+<script>
+	export let data;
+</script>
+
+<p>Page without load</p>
+
+<pre>{JSON.stringify(data)}</pre>