+++ packages/kit/test/prerendering/paths-base/test/test.js
@@ -10,6 +10,11 @@ const read = (file) => fs.readFileSync(`${build}/${file}`, 'utf-8');
 
 test('prerenders /path-base', () => {
 	const content = read('index.html');
+	assert.ok(content.includes('/path-base/favicon.png') && content.includes('/path-base/nested'));
+});
+
+test('prerenders /path-base/redirect', () => {
+	const content = read('redirect.html');
 	assert.equal(content, '<meta http-equiv="refresh" content="0;url=/path-base/dynamic/foo">');
 });