+++ packages/kit/types/internal.d.ts
@@ -271,6 +271,7 @@ export interface ServerMetadata {
 		{
 			prerender: PrerenderOption | undefined;
 			methods: HttpMethod[];
+			config: any;
 		}
 	>;
 }
@@ -311,6 +312,7 @@ export interface SSRNode {
 		ssr?: boolean;
 		csr?: boolean;
 		trailingSlash?: TrailingSlash;
+		config?: any;
 	};
 
 	server: {
@@ -320,6 +322,7 @@ export interface SSRNode {
 		csr?: boolean;
 		trailingSlash?: TrailingSlash;
 		actions?: Actions;
+		config?: any;
 	};
 
 	// store this in dev so we can print serialization errors
@@ -363,6 +366,7 @@ export interface PageNodeIndexes {
 export type SSREndpoint = Partial<Record<HttpMethod, RequestHandler>> & {
 	prerender?: PrerenderOption;
 	trailingSlash?: TrailingSlash;
+	config?: any;
 };
 
 export interface SSRRoute {