+++ packages/create-svelte/CHANGELOG.md
@@ -1,5 +1,13 @@
 # create-svelte
 
+## 2.3.3
+
+### Patch Changes
+
+- chore: bump `@sveltejs/kit` and `@sveltejs/adapter-auto` versions ([#8740](https://github.com/sveltejs/kit/pull/8740))
+
+- fix: adjust vite config type to work with vitest options ([#8871](https://github.com/sveltejs/kit/pull/8871))
+
 ## 2.3.2
 
 ### Patch Changes